api/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
openai
twilio
python-multipart
